Create Prime number
Write a java program to print "N" prime numbers within the range of X (lower limit) and Y
(upper limit). Prompt the values of X, Y, and N from the user. Ensure that the value Y is greater than
X and all the numbers are positive whole numbers. For any invalid input, ask for the input again.